repo-qa/requirements.txt
# Core dependencies
mini-swe-agent==1.17.3
tree-sitter==0.25.2
tree-sitter-python==0.25.0
networkx==3.4.2
litellm==1.81.5
python-dotenv>=1.0.0

# Configuration
pyyaml>=6.0

# Data analysis (optional, for later stages)
pandas>=2.0.0

# Testing
pytest>=7.0.0
